+++ deploy/models.py
@@ -2,7 +2,9 @@ from django.db import models
 from django.utils.translation import ugettext_lazy as _
 from django.utils import timezone
 from asset.models import Server
+from user.models.user import User
 
+__all__ = ['SaltMinion', 'Roster']
 
 class SaltMinion(models.Model):
     STATUS_CHOICE = (
@@ -31,5 +33,30 @@ class SaltMinion(models.Model):
 
     def __repr__(self):
         return self.hostname
+
+    __str__ = __repr__
+
+
+class Roster(models.Model):
+    STATUS_CHOICE = (
+        (0, 'Unknow'),
+        (1, 'Normal'),
+        (2, 'Deleted'),
+    )
+
+    uuid = models.CharField(_('UUID'), max_length=100, blank=True)
+    user = models.ForeignKey(User, blank=True, null=True)
+    file_name = models.CharField(_('Roster File Name'), max_length=200, blank=True)
+    upload_to = models.FileField(_('Upload to'), upload_to='roster/')
+    status = models.SmallIntegerField(_('Status'), choices=STATUS_CHOICE, default=0, blank=True)
+    create_time = models.DateTimeField(_('Create Time'), auto_now_add=True, null=True, blank=True)
+    update_time = models.DateTimeField(_('Update Time'), default=timezone.now(), null=True, blank=True)
+
+    class Meta:
+        db_table = 'deploy_roster'
+        ordering = ['file_name', ]
+
+    def __repr__(self):
+        return self.file_name
 
     __str__ = __repr__