+++ deploy/urls/views_urls.py
@@ -11,5 +11,6 @@ urlpatterns = [
     url(r'^minion-list/$', views.SaltMinionListView.as_view(), name='minion-list',
         kwargs={'path1': 'Minion List'}),
     url(r'^execute-command/$', views.ExecuteCommandView.as_view(), name='execute-command',
-        kwargs={'path1': 'Execute Command'})
+        kwargs={'path1': 'Execute Command'}),
+    url(r'^ssh/$', views.SaltSSHView.as_view(), name='ssh', kwargs={'path1': 'SSH'}),
 ]